Start to separate API specific code into separate packages
